A bit more refactoring
